+++ ui/Screen/Project/Source/Commits.svelte
@@ -1,12 +1,12 @@
 <script lang="typescript">
   import { selectCommit, store } from "../../../src/screen/project/source";
-  import { formatTime } from "../../../src/source";
   import type { CommitHeader } from "../../../src/source";
 
   import { Remote } from "../../../DesignSystem/Component";
-  import CommitTeaser from "../../../DesignSystem/Component/SourceBrowser/CommitTeaser.svelte";
 
-  const onSelect = (commit: CommitHeader) => {
+  import History from "../../../DesignSystem/Component/SourceBrowser/History.svelte";
+
+  const onSelect = ({ detail: commit }: { detail: CommitHeader }) => {
     selectCommit(commit);
   };
 </script>
@@ -18,56 +18,10 @@
     min-width: var(--content-min-width);
     padding: 2rem var(--content-padding) 0;
   }
-  .commit-group header {
-    padding-bottom: 0.75rem;
-    padding-left: 1rem;
-    color: var(--color-foreground-level-6);
-  }
-  .commit-group ul {
-    border: 1px solid var(--color-foreground-level-3);
-    border-radius: 0.25rem;
-    margin-bottom: 2rem;
-  }
-  .commit {
-    border-bottom: 1px solid var(--color-foreground-level-3);
-    cursor: pointer;
-    display: block;
-    height: 3rem;
-    padding: 0.25rem 0;
-  }
-  .commit:first-child {
-    border-top-left-radius: 0.25rem;
-    border-top-right-radius: 0.25rem;
-  }
-  .commit:last-child {
-    border-bottom: none;
-    border-bottom-left-radius: 0.25rem;
-    border-bottom-right-radius: 0.25rem;
-  }
-  .commit:hover {
-    background: var(--color-foreground-level-1);
-  }
 </style>
 
 <div class="commits-page" data-cy="commits-page">
   <Remote {store} let:data={{ history }}>
-    {#each history.history as group (group.time)}
-      <div class="commit-group">
-        <header>
-          <p>{formatTime(group.time * 1000)}</p>
-        </header>
-        <ul>
-          {#each group.commits as commit (commit.sha1)}
-            <li class="commit" on:click={() => onSelect(commit)}>
-              <CommitTeaser
-                {commit}
-                style="background: none; --commit-message-color:
-                var(--color-foreground-level-6); --commit-sha-color:
-                var(--color-foreground)" />
-            </li>
-          {/each}
-        </ul>
-      </div>
-    {/each}
+    <History {history} on:seelct={onSelect} />
   </Remote>
 </div>
